Refactoring test code
