--- Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import FABoLLSlidableTabBarTests
-
-var tests = [XCTestCaseEntry]()
-tests += FABoLLSlidableTabBarTests.allTests()
-XCTMain(tests)
